Update patch for ax_c_float_words_bigendian.m4
The original patch:
https://github.com/autoconf-archive/autoconf-archive/pull/279
broke endianness detection on Emscripten. This was fixed in this followup:
https://github.com/autoconf-archive/autoconf-archive/pull/316

We also have worked around the problem by adjusting `configure.ac` in the Python
repo to handle this case:
https://github.com/python/cpython/pull/126387
But it's useful for us to use the upstream autoconf-archive detection so that we
can be completely sure that the fix works in practice.